use ftrack server from ftrack common
