+++ services/processor/processor/__init__.py
@@ -1,4 +1,4 @@
-from .ftrack_server import main
+from .server import main
 
 
 __all__ = (
