+++ services/processor/processor/__main__.py
@@ -1,4 +1,4 @@
-from .ftrack_server import main
+from .server import main
 
 
 if __name__ == "__main__":
